perf(test): consolidate TUI PTY session coverage
