+++ src/tui/tui-session-identity-pty.e2e.test.ts
@@ -109,15 +109,16 @@ it("hides a stale approval when startup restores the remembered session", async 
   }
 }, 65_000);
 
-it("restores a remembered global session before startup admits the first send", async () => {
-  const stateDir = tempDirs.make("openclaw-tui-global-session-");
-  const marker = "global startup session proof";
+it("restores a remembered global session while keeping pre-ready input editable", async () => {
+  const stateDir = tempDirs.make("openclaw-tui-startup-session-");
+  const marker = "startup remembered session proof";
   await seedRememberedSession(stateDir, "global");
   const fixture = await startTuiFixture({
     env: {
       OPENCLAW_STATE_DIR: stateDir,
       OPENCLAW_TUI_PTY_PICKER_FIXTURE: "1",
       OPENCLAW_TUI_PTY_PICKER_SESSION_KEY: "global",
+      OPENCLAW_TUI_PTY_RESTORE_DELAY_MS: "400",
     },
   });
 
@@ -132,43 +133,6 @@ it("restores a remembered global session before startup admits the first send", 
       includeUnknown: false,
       agentId: "main",
     });
-
-    await fixture.run.waitForOutput("local ready", STARTUP_TIMEOUT_MS);
-    await fixture.run.write(`${marker}\r`, { delay: false });
-    await fixture.waitForLogEntry(
-      (entry) => entry.method === "sendChat" && objectFieldEquals(entry, "message", marker),
-      STARTUP_TIMEOUT_MS,
-    );
-
-    const sends = (await readFixtureLog(fixture.logPath)).filter(
-      (entry) => entry.method === "sendChat",
-    );
-    expect(sends).toHaveLength(1);
-    expect(sends[0]?.payload).toMatchObject({ sessionKey: "global", agentId: "main" });
-  } finally {
-    await fixture.cleanup();
-  }
-}, 65_000);
-
-it("keeps startup input editable until the remembered session and history are stable", async () => {
-  const stateDir = tempDirs.make("openclaw-tui-startup-session-");
-  const marker = "startup remembered session proof";
-  await seedRememberedSession(stateDir);
-  const fixture = await startTuiFixture({
-    env: {
-      OPENCLAW_STATE_DIR: stateDir,
-      OPENCLAW_TUI_PTY_PICKER_FIXTURE: "1",
-      OPENCLAW_TUI_PTY_RESTORE_DELAY_MS: "400",
-    },
-  });
-
-  try {
-    await fixture.waitForLogEntry(
-      (entry) =>
-        entry.method === "listSessions" &&
-        objectFieldEquals(entry, "search", REMEMBERED_SESSION_KEY),
-      STARTUP_TIMEOUT_MS,
-    );
     const outputOffset = fixture.run.visibleOutput().length;
     await fixture.run.write(`${marker}\r`, { delay: false });
     const decision = await waitForSubmitDecision({ fixture, marker, outputOffset });
@@ -177,7 +141,7 @@ it("keeps startup input editable until the remembered session and history are st
     const rows = await waitForSynchronizedFrameRows(
       fixture.run,
       (frame) =>
-        frame.some((row) => row.includes("session picker-target")) &&
+        frame.some((row) => row.includes("session global")) &&
         frame.some((row) => row.includes("local ready")) &&
         frame.some((row) => row.includes(marker)),
       STARTUP_TIMEOUT_MS,
@@ -190,7 +154,7 @@ it("keeps startup input editable until the remembered session and history are st
       (entry) => entry.method === "sendChat" && objectFieldEquals(entry, "message", marker),
       STARTUP_TIMEOUT_MS,
     );
-    expect(sent.payload).toMatchObject({ sessionKey: REMEMBERED_SESSION_KEY });
+    expect(sent.payload).toMatchObject({ sessionKey: "global", agentId: "main" });
     expect(markerSends(await readFixtureLog(fixture.logPath), marker)).toHaveLength(1);
   } finally {
     await fixture.cleanup();
